refactor(agent-lane): 阶段 4 前置 ③ · 接缝类型搬家，harness/runtime 零反向引用
阶段 4 的切换 PR 要整个删掉 `electron/harness/runtime/`。今天挡在前面的是一批
反向边：20+ 个活过切换的文件从 `harness/runtime/runtimePort.ts` 拿类型，
`electron/agentLane/` 从 `harness/runtime/pi/model.mts` 拿 `createNomiProvider`
——新通路依赖着一个要删的目录。本 PR 只搬家，零行为变化。

搬了什么 → 去哪：
- NomiModelConfig / NomiPricingBasis → electron/shared/agentLane/laneModelConfig.ts
- RuntimeToolDescriptor/ToolCall/ToolDecision/ToolCallRecord/Usage/FinishReason/
  ErrorFacts/ActivityEvent → electron/shared/agentCapabilities/transportContracts.ts
- observeStream.mts → electron/agentLane/laneStreamObserver.mts
- providerGuard.mts → electron/agentLane/laneProviderGuard.mts
- model.mts 的 modelConfigSchema/protocols/anthropicFetch/modelCost/createNomiProvider
  → electron/agentLane/laneModelProvider.mts

两条经裁决的偏离：
1. runtimePort.ts 不做「纯 re-export」。RuntimeTurnRequest/Hooks/Result/
   RunAgentTurn/SnapshotCodec 依赖 harness/context/promptPipe，且没有一个消费者
   活过切换 PR——它们留在原地随目录一起死。硬搬进 shared 会造一条 shared→harness
   的新反向边，比不搬更糟。只搬有存活消费者的那批。
2. 判据是「import 语句为零」而不是文本 grep 为零。剩下的 11 处文本命中全是注释、
   pi 包自己的内部路径（laneTools.mts:211）、以及影子期故意断言旧路径的
   laneShadowStructure.test.ts:28。

旧通路一行逻辑不改：pi/observeStream.mts 留成纯 re-export（不是第二份实现，P1），
所以 run.mts 与 errorFacts.mts 的 diff 为零；agentChatV2.ts 只拆了一条 import。

验证：typecheck ✅ / vitest 1282 files ✅ / test:agent-runtime 290-290 ✅ /
check:boundaries 基线 80 未变 ✅ / check:framework-surface ✅ / check:filesize ✅ /
lint:ci 81 warnings（棘轮 98）✅
