+++ electron/capabilityCore/generationTransportAdapters.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import type { RuntimeToolCall, RuntimeToolDecision } from "../harness/runtime/runtimePort";
+import type { RuntimeToolCall, RuntimeToolDecision } from "../shared/agentCapabilities/transportContracts";
 import {
   generationPlanInputSchema,
   generationStatusInputSchema,
