+++ electron/capabilityCore/skillWriteTransportAdapters.ts
@@ -1,6 +1,6 @@
 import { createHash } from "node:crypto";
 
-import type { RuntimeToolCall, RuntimeToolDecision } from "../harness/runtime/runtimePort";
+import type { RuntimeToolCall, RuntimeToolDecision } from "../shared/agentCapabilities/transportContracts";
 import {
   SKILL_WRITE_CAPABILITY,
   skillWriteInputForAlias,
