+++ scripts/check-network-entry.mjs
@@ -8,7 +8,7 @@ const root = fileURLToPath(new URL('../', import.meta.url));
 const EXPLICIT_BOUNDARIES = new Set([
   'electron/appFetch.ts', // injects the app-owned dispatcher into native fetch
   'electron/harness/runtime/pi/run.mts', // isolated SDK adapter; host always injects appFetch
-  'electron/harness/runtime/pi/model.mts', // standalone SDK model tests; normal runtime injects fetch
+  'electron/agentLane/laneModelProvider.mts', // standalone SDK model tests; normal runtime injects fetch
   'electron/capabilityCore/mcpNodeLauncher.ts', // pure Node CLI -> authenticated localhost RPC only
 ]);
 const CLIENT_METHODS = new Map([
