+++ tests/agent-runtime/errorFacts.test.mts
@@ -1,7 +1,8 @@
 import assert from 'node:assert/strict';
 import { setImmediate } from 'node:timers/promises';
 import { test } from 'node:test';
-import type { NomiModelConfig, RuntimeErrorFacts } from '../../electron/harness/runtime/runtimePort.js';
+import type { NomiModelConfig } from '../../electron/shared/agentLane/laneModelConfig.js';
+import type { RuntimeErrorFacts } from '../../electron/shared/agentCapabilities/transportContracts.js';
 import { createErrorFacts } from '../../electron/harness/runtime/pi/errorFacts.mjs';
 
 const model: NomiModelConfig = { kind: 'openai-compatible', providerId: 'nomi-review', modelId: 'review-model',
