+++ tests/agent-runtime/lane-shadow-parity.test.mts
@@ -18,7 +18,8 @@ import { z } from 'zod';
 import { openLane } from '../../electron/agentLane/laneHost.mjs';
 import { createDocumentLaneTools } from '../../electron/agentLane/laneDocumentTools.js';
 import { runAgentTurn } from '../../electron/harness/runtime/pi/nativeLoader.cjs';
-import type { RuntimeActivityEvent, RuntimeTurnRequest } from '../../electron/harness/runtime/runtimePort.js';
+import type { RuntimeTurnRequest } from '../../electron/harness/runtime/runtimePort.js';
+import type { RuntimeActivityEvent } from '../../electron/shared/agentCapabilities/transportContracts.js';
 import { createHttpFixture, type FixtureReply } from './httpFixture.mjs';
 import { createDocumentPort, LANE_SYSTEM_PROMPT } from './laneFixture.mjs';
 import type { LaneProjection } from '../../electron/shared/agentLane/laneContracts.js';
