+++ tests/agent-runtime/runtime-cancel.test.mts
@@ -2,7 +2,7 @@ import assert from 'node:assert/strict';
 import { setImmediate } from 'node:timers/promises';
 import { test } from 'node:test';
 import { z } from 'zod';
-import type { RuntimeActivityEvent, RuntimeToolDecision } from '../../electron/harness/runtime/runtimePort.js';
+import type { RuntimeActivityEvent, RuntimeToolDecision } from '../../electron/shared/agentCapabilities/transportContracts.js';
 import { runAgentTurn } from '../../electron/harness/runtime/pi/nativeLoader.cjs';
 import { createRuntimeFixture, type FixtureReply } from './httpFixture.mjs';
